refactor: name issue finders by purpose
