+++ find_pattern_coverage_issue.py
